Add JSON parsing of tags
Added a system to manage loading/saving of tags. Creation/deletion is still pending, but shouldn't be hard to add.

There seems to be an issue with loading the UI tabs, I'm not sure if it's caused by code within this extension or some weird caching issue with Gradio or A1111.

Additional changes:
- Cleaned up some code and extracted (mostly) all paths into it's own file
- Fixed .gitignore (you should never be ignoring everything)

NOTE: My changes are simply added on the side of everything else as of now, this PR is not merge ready.