Experimental dataframe layout
Seeing as elements with labels are cached, using a dataframe seems to be the most practical path forward.
There are still a few issues with this approach however. Mainly the fact that you can't seem to edit the row_count property of a dataframe programmatically (at least I couldn't figure out how)